refactor(deferred): extract deferred event handlers to sibling module
Move read_deferred_node, handle_deferred_register, handle_deferred_evaluate
and their offset constants out of sys_event_handlers.rs into a new
fast/basic_block/deferred_handlers.rs. Keeps the deferred glue self-contained
and lets the parent dispatch import them by name.

Addresses self-review comment 3239883343.
